+++ contracts/p0/Broker.sol
@@ -19,13 +19,13 @@ contract BrokerP0 is ComponentP0, IBroker {
 
     // The fraction of the supply of the bidding token that is the min bid size in case of default
     uint192 public constant MIN_BID_SHARE_OF_TOTAL_SUPPLY = 1e9; // (1} = 1e-7%
-    uint32 public constant MAX_AUCTION_LENGTH = 604800; // {s} max valid duration -1 week
+    uint48 public constant MAX_AUCTION_LENGTH = 604800; // {s} max valid duration -1 week
 
     IGnosis public gnosis;
 
     mapping(address => bool) private trades;
 
-    uint32 public auctionLength; // {s} the length of an auction
+    uint48 public auctionLength; // {s} the length of an auction
 
     bool public disabled;
 
@@ -33,7 +33,7 @@ contract BrokerP0 is ComponentP0, IBroker {
         IMain main_,
         IGnosis gnosis_,
         ITrade, // Added for Interface compatibility with P1
-        uint32 auctionLength_
+        uint48 auctionLength_
     ) public initializer {
         __Component_init(main_);
         gnosis = gnosis_;
@@ -75,7 +75,7 @@ contract BrokerP0 is ComponentP0, IBroker {
     // === Setters ===
 
     /// @custom:governance
-    function setAuctionLength(uint32 newAuctionLength) public governance {
+    function setAuctionLength(uint48 newAuctionLength) public governance {
         require(
             newAuctionLength > 0 && newAuctionLength <= MAX_AUCTION_LENGTH,
             "invalid auctionLength"
