+++ contracts/p0/Main.sol
@@ -22,8 +22,8 @@ contract MainP0 is Initializable, Auth, ComponentRegistry, IMain {
     function init(
         Components memory components,
         IERC20 rsr_,
-        uint32 shortFreeze_,
-        uint32 longFreeze_
+        uint48 shortFreeze_,
+        uint48 longFreeze_
     ) public virtual initializer {
         __Auth_init(shortFreeze_, longFreeze_);
         __ComponentRegistry_init(components);
