+++ contracts/p1/BackingManager.sol
@@ -20,15 +20,15 @@ contract BackingManagerP1 is TradingP1, IBackingManager {
     using FixLib for uint192;
     using SafeERC20Upgradeable for IERC20Upgradeable;
 
-    uint32 public constant MAX_TRADING_DELAY = 31536000; // {s} 1 year
+    uint48 public constant MAX_TRADING_DELAY = 31536000; // {s} 1 year
     uint192 public constant MAX_BACKING_BUFFER = 1e18; // {%}
 
-    uint32 public tradingDelay; // {s} how long to wait until resuming trading after switching
+    uint48 public tradingDelay; // {s} how long to wait until resuming trading after switching
     uint192 public backingBuffer; // {%} how much extra backing collateral to keep
 
     function init(
         IMain main_,
-        uint32 tradingDelay_,
+        uint48 tradingDelay_,
         uint192 backingBuffer_,
         uint192 maxTradeSlippage_
     ) external initializer {
@@ -178,7 +178,7 @@ contract BackingManagerP1 is TradingP1, IBackingManager {
     // === Setters ===
 
     /// @custom:governance
-    function setTradingDelay(uint32 val) public governance {
+    function setTradingDelay(uint48 val) public governance {
         require(val <= MAX_TRADING_DELAY, "invalid tradingDelay");
         emit TradingDelaySet(tradingDelay, val);
         tradingDelay = val;
