+++ contracts/p1/Main.sol
@@ -33,8 +33,8 @@ contract MainP1 is
     function init(
         Components memory components,
         IERC20 rsr_,
-        uint32 shortFreeze_,
-        uint32 longFreeze_
+        uint48 shortFreeze_,
+        uint48 longFreeze_
     ) public virtual initializer {
         __Auth_init(shortFreeze_, longFreeze_);
         __ComponentRegistry_init(components);
